food added

+++ src/utils/helpers/files/index.ts
@@ -3,7 +3,7 @@ import fs from 'fs'
 const saveFile = (file: Express.Multer.File): string => {
   if(file && file.mimetype.includes('image')) {
     const parts = file.originalname.split('.')
-    const fileName = parts[0] + '-' + Date.now() + '.' + parts[parts.length - 1]
+    const fileName = parts[0].split(' ').join('_') + '-' + Date.now() + '.' + parts[parts.length - 1]
     const targetPath = 'static/' + fileName
     fs.writeFileSync(targetPath, file.buffer)
   
@@ -14,8 +14,12 @@ const saveFile = (file: Express.Multer.File): string => {
 }
 
 const deleteFile = (fileName: string) => {
-  const targetPath = 'static/' + fileName
-  fs.unlinkSync(targetPath)
+  if(fileName) {
+    const targetPath = 'static/' + fileName
+    if(fs.existsSync(targetPath)) {
+      fs.unlinkSync(targetPath)
+    }
+  }
 }
 
 export default {
